requirements.txt
PyYAML>=6.0.1
qrcode>=7.4.2
requests>=2.31.0
rich>=13.7.0